+++ src/info.c
@@ -76,7 +76,7 @@ static int           info_lastTab; /**< Last open tab. */
 
 static CstSlotWidget info_eq;
 static CstSlotWidget info_eq_weaps;
-static int          *info_factions;
+static int64_t      *info_factions;
 
 static int selectedMission = 0; /**< Current index in the missions list-box. */
 static int selectedLog     = 0;
@@ -1092,7 +1092,7 @@ static void cargo_update( unsigned int wid, const char *str )
       l += scnprintf( &desc[l], sizeof( desc ) - l, "\n\n%s",
                       _( "Illegalized by the following factions:\n" ) );
       for ( int i = 0; i < array_size( com->illegalto ); i++ ) {
-         int f = com->illegalto[i];
+         int64_t f = com->illegalto[i];
          if ( !faction_isKnown( f ) )
             continue;
 
@@ -1224,8 +1224,8 @@ static void standings_close( unsigned int wid, const char *str )
 
 static int factionsSort( const void *p1, const void *p2 )
 {
-   int    f1, f2;
-   double v1, v2;
+   int64_t f1, f2;
+   double  v1, v2;
    f1 = *(int *)p1;
    f2 = *(int *)p2;
    v1 = round( faction_reputation( f1 ) );
@@ -1298,7 +1298,7 @@ static void standings_update( unsigned int wid, const char *str )
    const glTexture *t;
    int              w, h, lw, l;
    double           m, ml;
-   const int       *flist;
+   const int64_t   *flist;
    char             buf[STRMAX];
 
    /* Get dimensions. */
@@ -1354,7 +1354,7 @@ static void standings_update( unsigned int wid, const char *str )
    if ( array_size( flist ) > 0 ) {
       int added = 0;
       for ( int i = 0; i < array_size( flist ); i++ ) {
-         int f = flist[i];
+         int64_t f = flist[i];
          if ( faction_isStatic( f ) || !faction_isKnown( f ) ||
               faction_isInvisible( f ) || faction_isDynamic( f ) )
             continue;
@@ -1372,7 +1372,7 @@ static void standings_update( unsigned int wid, const char *str )
    if ( array_size( flist ) > 0 ) {
       int added = 0;
       for ( int i = 0; i < array_size( flist ); i++ ) {
-         int f = flist[i];
+         int64_t f = flist[i];
          if ( faction_isStatic( f ) || !faction_isKnown( f ) ||
               faction_isInvisible( f ) || faction_isDynamic( f ) )
             continue;
